backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
faiss-cpu==1.7.4
openai==1.6.1
python-dotenv==1.0.0
httpx==0.25.2
pypdf2==3.0.1
tiktoken==0.5.2
aiofiles==23.2.1
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
